+++ .travis.yml
@@ -8,6 +8,6 @@ install:
     - python setup.py install
     - pip install pylint pytest coverage
 script:
-    - coverage run --source linode_api -m pytest
+    - coverage run --source linode_api4 -m pytest
     - coverage report
-    - pylint linode
+    - pylint linode_api4
