+++ docs/index.rst
@@ -1,5 +1,5 @@
-linode-api
-==========
+linode_api4
+===========
 
 .. note::
    These docs are currently in development, and are therefore incomplete.
@@ -8,7 +8,7 @@ linode-api
    .. _Official API Documentation: https://developers.linode.com/api/v4
 
 This is the documentation for the official Python bindings of the Linode
-API.  For API documentation, see `developers.linode.com`_.
+API v4.  For API documentation, see `developers.linode.com`_.
 
 This library can be used to interact with all features of the Linode API, and
 is compatible with Python 2 and 3.
@@ -20,12 +20,12 @@ Installation
 
 To install through pypi::
 
-   pip install linode-api
+   pip install linode_api4
 
 To install from source::
 
-  git clone https://github.com/linode/linode-api-python
-  cd python-linode-api
+  git clone https://github.com/linode/linode_api4
+  cd linode_api4
   python setup.py install
 
 For more information, see our :doc:`Getting Started<guides/getting_started>`
@@ -40,7 +40,7 @@ Table of Contents
    guides/getting_started
    guides/core_concepts
    guides/oauth
-   linode_api/linode_client
-   linode_api/login_client
-   linode_api/paginated_list
-   linode_api/objects/filtering
+   linode_api4/linode_client
+   linode_api4/login_client
+   linode_api4/paginated_list
+   linode_api4/objects/filtering
