+++ examples/install-on-linode/app.py
@@ -1,8 +1,8 @@
 import re
 from flask import Flask, redirect, request, render_template, session, send_from_directory
 from flask.ext.session import Session
-from linode_api import (LinodeClient, LinodeLoginClient, StackScript, Image, Region,
-                        Type, OAuthScopes)
+from linode_api4 import (LinodeClient, LinodeLoginClient, StackScript, Image, Region,
+                         Type, OAuthScopes)
 import config
 
 app=Flask(__name__)
