+++ examples/install-on-linode/requirements.txt
@@ -1,3 +1,3 @@
-linode-api
+linode_api4
 Flask
 Flask-Session
