--- linode_api/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import absolute_import # python2 imports should be absolute
-
-from linode_api.objects import *
-from linode_api.errors import ApiError, UnexpectedResponseError
-from linode_api.linode_client import LinodeClient
-from linode_api.login_client import LinodeLoginClient, OAuthScopes
-from linode_api.paginated_list import PaginatedList
