+++ linode_api4/linode_client.py
@@ -7,14 +7,14 @@ from datetime import datetime
 import pkg_resources
 import requests
 
-from linode_api.errors import ApiError, UnexpectedResponseError
-from linode_api.objects import *
-from linode_api.objects.filtering import Filter
+from linode_api4.errors import ApiError, UnexpectedResponseError
+from linode_api4.objects import *
+from linode_api4.objects.filtering import Filter
 
 from .common import load_and_validate_keys
 from .paginated_list import PaginatedList
 
-package_version = pkg_resources.require("linode-api")[0].version
+package_version = pkg_resources.require("linode_api4")[0].version
 
 logger = logging.getLogger(__name__)
 
@@ -753,7 +753,7 @@ class LinodeClient:
 
     @property
     def _user_agent(self):
-        return '{}python-linode-api/{} {}'.format(
+        return '{}python-linode_api4/{} {}'.format(
                 '{} '.format(self._add_user_agent) if self._add_user_agent else '',
                 package_version,
                 requests.utils.default_user_agent()
