+++ linode_api4/objects/nodebalancer.py
@@ -2,9 +2,9 @@ from __future__ import absolute_import
 
 import os
 
-from linode_api.errors import UnexpectedResponseError
-from linode_api.objects import Base, DerivedBase, Property, Region
-from linode_api.objects.networking import IPAddress
+from linode_api4.errors import UnexpectedResponseError
+from linode_api4.objects import Base, DerivedBase, Property, Region
+from linode_api4.objects.networking import IPAddress
 
 
 class NodeBalancerNode(DerivedBase):
