+++ test/base.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from mock import patch
 
-from linode_api import LinodeClient
+from linode_api4 import LinodeClient
 
 from .fixtures import TestFixtures
 
@@ -72,7 +72,7 @@ class MethodMock:
         Begins the method mocking
         """
         self.patch = patch(
-            'linode_api.linode_client.requests.'+self.method,
+            'linode_api4.linode_client.requests.'+self.method,
             return_value=MockResponse(200, self.return_dct)
         )
         self.mock = self.patch.start()
@@ -130,7 +130,7 @@ class ClientBaseCase(TestCase):
     def setUp(self):
         self.client = LinodeClient('testing', base_url='/')
 
-        self.get_patch = patch('linode_api.linode_client.requests.get',
+        self.get_patch = patch('linode_api4.linode_client.requests.get',
                 side_effect=mock_get)
         self.get_patch.start()
