+++ test/objects/longview_test.py
@@ -1,8 +1,8 @@
 from datetime import datetime
 from test.base import ClientBaseCase
 
-from linode_api.objects import LongviewClient, LongviewSubscription
-from linode_api.objects.base import MappedObject
+from linode_api4.objects import LongviewClient, LongviewSubscription
+from linode_api4.objects.base import MappedObject
 
 
 class LongviewClientTest(ClientBaseCase):
