+++ test/objects/volume_test.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from test.base import ClientBaseCase
 
-from linode_api.objects import Volume
+from linode_api4.objects import Volume
 
 
 class VolumeTest(ClientBaseCase):
