+++ tox.ini
@@ -10,6 +10,6 @@ deps =
     pylint
 commands =
     python setup.py install
-    coverage run --source linode_api -m pytest
+    coverage run --source linode_api4 -m pytest
     coverage report
-    pylint linode_api
+    pylint linode_api4
